Updating the version # in the files to 1.0.1_dev.
Outputting the link outputted in PROFILE_BLOG in U_VIEW_BLOG (PROFILE_BLOG was just added recently).
Along with that adding an option in the ACP to disable outputting the link and text to the custom profile fields (so it is easier to alter the output for templates).

Fixing a hole where users could rate items higher or lesser than the max/min allowed rating.

Fixing the alt text on rating images.

git-svn-id: http://userblogmod.googlecode.com/svn/trunk@393 a294dfe0-ea36-0410-9d4a-05498620a423