Upping the version to 1.0.2_dev.
Give the option to style authors to use the plugins system for their style.
This should not be used as a normal plugin, but mostly to alter outputted data to fit the way the style author wants it to.
For example, the style author may want to force the date/time outputted to a specific format, etc for this specific style.

git-svn-id: http://userblogmod.googlecode.com/svn/trunk@420 a294dfe0-ea36-0410-9d4a-05498620a423